refactor : main ->  QuizList 역할 분리
