+++ styles/common.ts
@@ -1,16 +1,18 @@
 import styled from 'styled-components';
 
-export const MainBtn = styled.button`
+export const MainBtn = styled.button<{ isBlur?: boolean }>`
   display: flex;
   align-items: center;
   justify-content: center;
   font-size: ${({ theme }) => theme.fontSize.body_1};
   font-weight: ${({ theme }) => theme.fontWeight.bold};
-  padding: 12px 100px;
+  padding: 12px 0;
+  width: 100%;
   border: none;
   color: #fff;
   border-radius: 8px;
   border: 1px solid #0500c9;
+  opacity: ${({ isBlur }) => (isBlur ? 0.5 : 1)};
   background: #3b27ff;
   box-shadow: 0px 4px 0px 0px #0500c9;
   word-break: keep-all;
@@ -30,6 +32,7 @@ export const Sketchbook = styled.div`
   border-radius: 8px;
   border: 0.75px solid #00fba5;
   background: #fff;
+  padding: 0 4%;
   box-shadow: 1px -2px 0px 0px #00fba5;
   &::before {
     content: '';
